Change `is_joined` to `when` on `LinkResult`
LinkResult now sends the join date of the member instead of checking
if the user has more then a week server-side.

This check was dropped as it was overly complex and
counter-intuitive and splitted users in those that were trusted (Had
more then a week on the guild) and those that weren't.

This system may get re-added if need be but it isn't a requirement for
the bakend service to function.