Add `Res<T>` type alias
This type alias is the Result type for routes so that error handling
is standardized between all routes.

Main rationale for that is that prior to this change all routes used
HTTP Status Codes to send which error happened, and thats why every
error path used a different status code, despite InternalServerError
being a more valid response for most of them.

All information about where and what failed was lost to the API
consumers which is bad for debugging if something happened.

I'd had to log which StatusCode happened and when it happened at the
call site, and that would be ugly and annoying to discern what
happened on the server-side.

Now the server shall return a well formed error with a correct
StatusCode and propagate the cause of that error, if it exists.