Change `ValidIp(Uuid)` to `ValidIp(Allowlist)`
This change was done in order for callers to be able to know which allowlist entry
was matched/updated, without needing to call the database again.

This should simplify most of the code which does CIDR checking, since calling
`self.check_cidr` should suffice for most of the cases where you only need to
know which entry of {Allow,Black}list was matched.